// kservices_ts_api/src/api/domain/entities/sme.entity.ts
export type callListFetchRequest = {
  id: number;
  type: string;
  isDownload: string;
  initialRecord: number;
  batchSize: number;
  startDate: string;
  startDate_op: string;
  endDate: string;
  endDate_op: string;
  callDirectionStatus: string;
  callDirectionStatus_op: string;
  duration: string;
  duration_op: string;
  calledNumber: string;
  calledNumber_op: string;
  callingNumber: string;
  callingNumber_op: string;
  agentName: string;
  agentName_op: string;
  callStatus: string;
  callStatus_op: string;
  answerStatus: string;
  answerStatus_op: string;
  remarks: string;
  remarks_op: string;
  callId: string;
  callId_op: string;
  callFlow: string,
  callFlow_op: string,
};
// ...
export type fetchAgentReport = {
  sme_id: number;
  isDownload: string;
  initialRecord: number;
  batchSize: number;
  startDate: string;
  startDate_op: string;
  endDate: string;
  endDate_op: string;
  duration: string;
  duration_op: string;
  customerAni: string;
  customerAni_op: string;
  agentNumber: string;
  agentNumber_op: string;
  agentName: string;
  agentName_op: string;
  responseMessage: string;
  responseMessage_op: string;
  callMode: string;
  callMode_op: string;
  sessionId: string;
  sessionId_op: string;
  orderBy: string;
  orderBy_op: string;
  status: string;
  status_op: string;
};